LPS-57770 Move out portal cache cluster link replication logic as an osgi bundle
