+++ modules/portal-cache/portal-cache-cluster/src/com/liferay/portal/cache/cluster/internal/PortalCacheClusterEvent.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.kernel.cache.cluster;
+package com.liferay.portal.cache.cluster.internal;
 
 import com.liferay.portal.kernel.cache.PortalCache;
 import com.liferay.portal.kernel.util.StringBundler;
