+++ modules/portal-cache/portal-cache-cluster/src/com/liferay/portal/cache/cluster/internal/bootstrap/ClusterLinkBootstrapLoaderHelperUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.kernel.cache.bootstrap;
+package com.liferay.portal.cache.cluster.internal.bootstrap;
 
 import com.liferay.portal.kernel.cache.PortalCache;
 import com.liferay.portal.kernel.cache.PortalCacheHelperUtil;
